Moved thread class to another file
